기존 rds -> rds_etc, rds_kakao로 분리
